requirements.txt
python-binance==1.0.19
requests>=2.25.1
pandas>=1.3.0
colorama>=0.4.4
python-dotenv>=0.19.0